webhook: allow rendering app on host display
When workloads are annotated with `wavy.squat.ai/host=true`, Wavy
renders their applications on a physical device connected to the host,
for example a monitor connected to a server. This mode of operation
allows nodes in a Kubernetes cluster to serve as display kiosks
rendering an application on a display.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>