webhook,README: add support for X
Currently, Wavy only supports workloads that run exclusively on Wayland,
because although Xwayland is installed in the Sway container, the X
socket is not available to the workload. This commit adds support for X
in workloads by sharing the X socket through a VolumeMount. Support for
X can also be disbled in workloads using the new `wavy.squat.ai/x`
annotation. Finally, this commit changes the README to deploy Signal
Desktop rather than Inkscape because the deployment of Inkscape was
somewhat confusing due to the installation and tricky readiness check.
The new example also asserts support for X. This current example has its
own downside, which is that Signal's config directory has to be
explicitly mounted into the container at `/nonexistent/.config`, which
is odd and might be confusing to users. A simpler example would be
better. A future enhancement could be to create a directory of example
applications.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>