e2e: make tests more robust
Currently the e2e tests will ocasionally fail when a request is made to
the webhook before the CA has been injected into the webhook
configuration. This commit fixes the issue by ensuring that the job to
inject the CA has finished before patching the workload and thus
making a request to the webhook server.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>