x11: allow running as different user
This commit improves support for applications that run on X11.
Specifically, this commit allows the container running the X11
application to run as a different user than the container running
sway/Xwayland. This is accomplished by generating an .Xauthority file in
an init container and using that to authenticate with X.
This loosens the constraints for application developers/administrators
significantly and simplifies deployment, as well as finding applications
that "just work" with Wavy.

xref: https://github.com/wavyland/sway/pull/1

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>