e2e_test: make tests more complete
This commit enhances the e2e tests to verify that a GUI actually loads,
the VNC server works and can be scripted, and that doing so produces the
expected result. In this instance, we run a Signal container, open some
menus with vncdotool, capture the GUI with vnccapture, and compare that
the output matches an expected image.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>